// src/UniversalViewer/types.ts
export enum ColorFamily {
  NEUTRAL = 'neutral',
  RED = 'red',
  ORANGE = 'orange',
  YELLOW = 'yellow',
  GREEN = 'green',
  BLUE = 'blue',
  PURPLE = 'purple',
  PINK = 'pink',
  BROWN = 'brown',
  GREY = 'grey'
}

export interface RGBColor {
  r: number;
  g: number;
  b: number;
}
// ...
export interface CoreColorMetadata {
  hex: string;
  rgb: RGBColor;
  name?: string;
  family?: ColorFamily;
  tags?: string[];
}